chore: carry on even if migrations fail
It looks like the "mlflow db upgrade" command expects tables to already exist,
otherwise it fails with a "relation "metrics" does not exist" error.

However, we still want to run the upgrades later. So, we just march forward if
migrations have failed. Not _that_ great, because it would be good for thing to
error if they fail proper. However, for our case it's good enough.